+++ src/python/lib/strelkaSharedOptions.py
@@ -76,8 +76,8 @@ class StrelkaSharedWorkflowOptionsBase(ConfigureWorkflowOptions) :
                               "to estimate statistics from the input (such as expected depth per chromosome). "
                               "Only one BED file may be specified. (default: call the entire genome)")
         group.add_option("--runDir", type="string", metavar="DIR",
-                         help="Name of directory to be created where run scripts and output will be written. "
-                              "This directory must not already exist. (default: %default)")
+                         help="Name of directory to be created where all workflow scripts and output will be written. "
+                              "Each analysis requires a separate directory. (default: %default)")
 
 
     def addExtendedGroupOptions(self,group) :
@@ -139,6 +139,8 @@ class StrelkaSharedWorkflowOptionsBase(ConfigureWorkflowOptions) :
 
         statsMergeBin=joinFile(libexecDir,exeFile("MergeRunStats"))
 
+        workflowScriptName = "runWorkflow.py"
+
         # default memory request per process-type
         #
         # where different values are provided for SGE and local runs note:
@@ -190,8 +192,10 @@ class StrelkaSharedWorkflowOptionsBase(ConfigureWorkflowOptions) :
 
         assertOptionExists(options.runDir,"run directory")
         options.runDir=os.path.abspath(options.runDir)
-        if os.path.exists(options.runDir):
-            raise OptParseException("Targeted runDir already exists '%s'. Each analysis requires a separate runDir." % (options.runDir))
+
+        workflowScriptPath = os.path.join(options.runDir, options.workflowScriptName)
+        if os.path.exists(workflowScriptPath):
+            raise OptParseException("Run directory already contains workflow script file '%s'. Each analysis must be configured in a separate directory." % (workflowScriptPath))
 
         assertOptionExists(options.referenceFasta,"reference fasta file")
         options.referenceFasta=validateFixExistingFileArg(options.referenceFasta,"reference fasta file")